create a seperate calcUtils module to manager calculator features
